pyproject.toml
[project]
name = "immich-slideshow"
version = "1.0.0"
description = "Home Assistant integration for Immich photo slideshow with memories support"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "aioresponses>=0.7.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
